Refactoring chrome entry point (#1931)
* Refactoring chrome entry point

additional changes

Refactoring chrome entry point

* Update src/js/bootstrap.js
