+++ src/js/chrome.js
@@ -1,3 +1,6 @@
 import '../sass/chrome.scss';
 
+// Add Red Hat font to body
+document.querySelector('body').classList.add('pf-m-redhat-font');
+
 import('./bootstrap');
